Move the package classes that represent the built class to a new module
